+++ app/lib/storage.js
@@ -1,13 +1,22 @@
 const STORAGE = localStorage;
 
 export default class Storage {
-  static set(key, data) {
-    data = JSON.stringify(data);
+  static set(key, value) {
+    const data = typeof value === 'object' ?
+      JSON.stringify(value) :
+      value;
+
     STORAGE.setItem(key, data);
   }
 
   static get(key) {
     const data = STORAGE.getItem(key);
+
+    try {
+      JSON.parse(data);
+    } catch(e) {
+      return data;
+    }
 
     return JSON.parse(data);
   }