+++ app/sources/users.js
@@ -1,11 +1,11 @@
 import config from 'config';
-import http from 'lib/http';
+import { post } from 'j-fetch';
 
 export default class UsersSource {
   static urlRoot = `${config.apiTarget}/users`;
 
   static create(user) {
-    return http.post({ url: this.urlRoot, body: user })
+    return post({ url: this.urlRoot, body: user })
       .then(result => result.json());
   }
 }